UI Tests for CurrencyDropdown Widget
